requirements.txt
amqp==5.0.5
argon2-cffi==20.1.0
asgiref==3.3.1
backcall==0.2.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
cryptography==3.4.5
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.5
django-allauth==0.44.0
django-appconf==1.0.4
django-celery-beat==2.2.0
django-ckeditor==6.0.0
django-compressor==2.4
django-cors-headers==3.7.0
django-crispy-forms==1.11.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.1
django-formtools==2.2
django-haystack==3.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-reversion==3.0.9
django-simpleui==2021.3
django-thumbnail==11.13
django-timezone-field==4.1.1
django-vote==2.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
elasticsearch==2.0.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.10
importlib-metadata==3.4.0
ipython==7.16.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.18.0
kombu==5.0.2
MarkupPy==1.14
mysqlclient==2.0.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.14
pycparser==2.20
Pygments==2.7.4
PyJWT==1.7.1
PyMySQL==1.0.2
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.4.1
rcssmin==1.0.6
requests==2.25.1
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
tablib==3.0.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
